Removed Request core module, added RequestWrappers
The removal of the Request core module was triggered by the fact I was
getting rather annoyed at passing ResponseWriters and Requests around
and decided to attach ResponseWriters and Requests directly to the
Context object. Interface is still the same context.Request() and
context.ResponseWriter(), only now it has the permanent contract of being
directly on the context.

Wrappers are functions that are called with a context, and a
RequestHandler, the idea here, is that they can add code before and after
the request handler is called, they also return an ActionResult and can
skip calling the handler if they wish, so they can also be used to prevent
the execution of or change the result of a request handler.